Purchases work with stripe

+++ pages/api/product/[id].ts
@@ -7,6 +7,7 @@ import multiparty from "multiparty";
 import fs from 'fs'
 
 import { RequestBody } from "../product";
+import { deleteProduct, updateProduct } from "@/lib/utils/stripe";
 
 export const config = {
     api: {
@@ -35,7 +36,7 @@ export default async function handler(request: NextApiRequest, response: NextApi
                 id: id as string
             }
         })
-        if (deletedProducts) {
+        if (deletedProducts && await deleteProduct(id as string)) {
             fs.unlinkSync(`${UPLOAD_URL}/${deletedProducts.image.split('/')[2]}`)
             response.status(204).send(null)
         }
@@ -71,6 +72,7 @@ export default async function handler(request: NextApiRequest, response: NextApi
                 price: Number(data.fields.price[0]),
             }
         })
+        await updateProduct(updatedProduct)
         if (updatedProduct) response.status(200).send(updatedProduct)
         else response.status(404).send({ message: 'Product was not found! Make sure you delete an existing product.' })
     }
